–: Some reformat around json formatter
