--- lib/tools/json/json_feature_utils.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/widgets.dart';
-import 'package:provider/provider.dart';
-
-import 'feature/json_feature.dart';
-
-extension JsonFeatureContext on BuildContext {
-  JsonFeature jsonFeature(BuildContext context, {bool listen = false}) {
-    return Provider.of<JsonFeature>(context, listen: listen);
-  }
-}
